+++ rocAL/include/api/rocal_api_types.h
@@ -424,6 +424,20 @@ enum RocalOutOfBoundsPolicy {
     ROCAL_ERROR
 };
 
+/*! \brief rocAL MelScale formula enum
+ * \ingroup group_rocal_types
+ */
+enum RocalMelScaleFormula {
+    /*! \brief Slaney
+     * Follows Slaney’s MATLAB Auditory Modelling Work behavior
+     */
+    ROCAL_SLANEY = 0,
+    /*! \brief HTK
+     * Follows O’Shaughnessy’s book formula, consistent with Hidden Markov Toolkit(HTK), m = 2595 * log10(1 + (f/700))
+     */
+    ROCAL_HTK
+};
+
 /*! \brief Tensor Last Batch Policies
  *  \ingroup group_rocal_types
  */