+++ rocAL/include/augmentations/augmentations_nodes.h
@@ -65,3 +65,5 @@ THE SOFTWARE.
 #include "augmentations/arithmetic_augmentations/node_tensor_add_tensor.h"
 #include "augmentations/audio_augmentations/node_non_silent_region_detection.h"
 #include "augmentations/geometry_augmentations/node_slice.h"
+#include "augmentations/effects_augmentations/node_normalize.h"
+#include "augmentations/audio_augmentations/node_mel_filter_bank.h"