+++ rocAL_pybind/amd/rocal/types.py
@@ -108,6 +108,10 @@ from rocal_pybind.types import PAD
 from rocal_pybind.types import TRIMTOSHAPE
 from rocal_pybind.types import ERROR
 
+#     RocalMelScaleFormula
+from rocal_pybind.types import SLANEY
+from rocal_pybind.types import HTK
+
 #     RocalLastBatchPolicy
 from rocal_pybind.types import LAST_BATCH_FILL
 from rocal_pybind.types import LAST_BATCH_DROP
@@ -183,6 +187,9 @@ _known_types = {
 
     NTF: ("NTF", NTF),
     NFT: ("NFT", NFT),
+
+    SLANEY: ("SLANEY", SLANEY),
+    HTK: ("HTK", HTK),
 
     LAST_BATCH_FILL : ("LAST_BATCH_FILL", LAST_BATCH_FILL),
     LAST_BATCH_DROP : ("LAST_BATCH_DROP", LAST_BATCH_DROP),