+++ tests/cpp_api/audio_tests/audio_tests.py
@@ -34,7 +34,9 @@ test_case_augmentation_map = {
     6: "tensor_add_tensor",
     7: "tensor_mul_scalar",
     8: "non_silent_region",
-    9: "slice"
+    9: "slice",
+    10: "mel_filter_bank",
+    11: "normalize"
 }
 
 def run_unit_test(rocal_data_path, qa_mode, gpu, downmix, build_folder_path, case_list):